requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
mysql-connector-python==9.4.0
tensorflow==2.20.0
keras==3.11.3
numpy==2.2.6
pandas==2.3.0
scikit-learn==1.6.1
PyMuPDF==1.26.5
Werkzeug==3.1.3
python-docx==1.2.0
SQLAlchemy==2.0.44
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
requests==2.32.3
tqdm==4.67.1
h5py==3.14.0